feat(StackBuffer): Add StackBuffer
Restructure data layout to separate stack + height and type

+++ src/systems/PlayersSystem.c
@@ -7,7 +7,7 @@
 
 Players newPlayers( int const boardWidth )
 {
-    assert( ( boardWidth >= (int)BOARD_WIDTH_MIN ) && "Board width value too small" );
+    assert( ( boardWidth >= BOARD_WIDTH_MIN ) && "Board width value too small" );
     assert( ( boardWidth <= (int)BOARD_WIDTH_MAX ) && "Board width value too big" );
 
     Players players = { 0 };
